+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.3 (31.08.2018)
 - добавлены уровни для карт [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61)
 - для сущности `Cards` для роли `Organization` добавлен метод `getCardLevelList` [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61) 
+- для сущности `Cards` для роли `Organization` исправлена ошибка в постраничной навигации [issue#61](https://github.com/rarus/bonus-server-php-sdk/issues/61) 
 - множественные исправления в тестах
 
 ## 0.5.2 (27.08.2018)